Brings in upstream xgrammar v0.1.31 (mlc-ai/xgrammar main) while preserving
Rico's workarounds:
- Stag hash / fingerprint deduplication (structural_tag.cc/h)
- Regex excludes tests

Upstream changes incorporated:
- v0.1.31 bump and serialization v9
- RuleLevelCache (crossing-grammar cache, #526)
- Excludes for any_text and triggered_tags (#502, #515)
- Refactored json_schema_converter (#527) and structural_tag (#528)
- Multiple end token support (#504)
- Various bugfixes and CI improvements

Core compiler/functor files (grammar_compiler.cc, grammar_functor.cc/h)
taken from upstream since both sides implement crossing-grammar caching
with different APIs (CrossingCacheManager vs RuleLevelCache) and upstream's
version is the canonical 0.1.31 release.

+++ cpp/support/utils.h
@@ -25,7 +25,7 @@ namespace xgrammar {
  * \brief Hash and combine value into seed.
  * \ref https://www.boost.org/doc/libs/1_84_0/boost/intrusive/detail/hash_combine.hpp
  */
-inline void HashCombineBinary(size_t& seed, size_t value) {
+inline void HashCombineBinary(uint64_t& seed, uint64_t value) {
   seed ^= value + 0x9e3779b97f4a7c15ull + (seed << 6) + (seed >> 2);
 }
 
@@ -33,8 +33,8 @@ inline void HashCombineBinary(size_t& seed, size_t value) {
  * \brief Find the hash sum of several size_t args.
  */
 template <typename... Args>
-inline size_t HashCombine(Args... args) {
-  size_t seed = 0;
+inline uint64_t HashCombine(Args... args) {
+  uint64_t seed = 0;
   (..., HashCombineBinary(seed, args));
   return seed;
 }
